move lib/db/index -> config (as it mainly controls the sync config)
